(0.8.0) 2012-10-02 Roman - FIX:Archive: Single value, typical for 1h, request is fixed. - FIX:UI.Vision, UI.WebVision: Periods values processing for hours more is fixed.
git-svn-id: http://oscada.org/svn/trunk/OpenSCADA/@1925 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/ui/Vision/vis_shapes.h
@@ -359,7 +359,7 @@ class ShapeDiagram : public WdgShape
 		QColor		mColor;		//Values line color
 		char		mWidth;		//Line width in pixels
 		//> Archive
-		int		arh_per;	//Archive period
+		int64_t		arh_per;	//Archive period
 		int64_t		arh_beg;	//Archive begin time
 		int64_t		arh_end;	//Archive end time
 		//> Values
